try to resolve searchParamRegistry issue
